Updated rigor s3 class to support both boto and boto3. This modification prefers boto3 and falls back to boto.

+++ lib/s3.py
@@ -2,8 +2,19 @@
 
 from abc import ABCMeta, abstractmethod
 from io import BytesIO
-from boto.s3.connection import S3Connection
-from boto.s3.key import Key
+import sys
+try:
+	import boto3
+except ImportError:
+	try:
+		from boto.s3.connection import S3Connection
+		from boto.s3.key import Key
+	except ImportError:
+		print("ignore")
+
+
+import botocore
+
 
 class RigorS3Client(object):
 	"""
@@ -13,6 +24,7 @@ class RigorS3Client(object):
 	:type config: :py:class:`~rigor.config.RigorConfiguration`
 	:param str bucket: S3 bucket containing data
 	:param str credentials: name of credentials section
+	:param str session: the s3 session if one exists.
 	"""
 	__metaclass__ = ABCMeta
 
@@ -46,7 +58,7 @@ class RigorS3Client(object):
 
 	@abstractmethod
 	def delete(self, key):
-		"""
+		"""sudo 
 		Removes data at the given key from S3
 
 		:param str key: S3 key for the object to delete
@@ -75,11 +87,24 @@ class BotoS3Client(RigorS3Client):
 		if credentials:
 			connection_args.append(config.get(credentials, 'aws_access_key_id'))
 			connection_args.append(config.get(credentials, 'aws_secret_access_key'))
-		self._conn = S3Connection(*connection_args)
-		self.bucket = self._conn.get_bucket(bucket)
+		if 'boto3' in sys.modules:
+			if credentials:
+				session = boto3.session.Session(aws_access_key_id=config.get(credentials, 'aws_access_key_id'),
+													  aws_secret_access_key=config.get(credentials, 'aws_secret_access_key'))
+				self._conn = session.resource('s3')
+			else:
+				self._conn = boto3.resource('s3')
+
+			self.bucket = self._conn.Bucket(bucket)
+		else:
+			self._conn = S3Connection(*connection_args)
+			self.bucket = self._conn.get_bucket(bucket)
 
 	def get(self, key, local_file=None):
 		""" See :py:meth:`RigorS3Client.get` """
+		if 'boto3' in sys.modules:
+			return self.__boto3_get__(key, local_file)
+
 		fetched_key = self.bucket.get_key(key)
 		if fetched_key is None:
 			return None
@@ -93,6 +118,9 @@ class BotoS3Client(RigorS3Client):
 
 	def put(self, key, data):
 		""" See :py:meth:`RigorS3Client.put` """
+		if 'boto3' in sys.modules:
+			return self.__boto3_put__(key, data)
+
 		remote_key = Key(self.bucket)
 		remote_key.key = key
 		if hasattr(data, 'read'):
@@ -102,14 +130,66 @@ class BotoS3Client(RigorS3Client):
 
 	def delete(self, key):
 		""" See :py:meth:`RigorS3Client.delete` """
+		if 'boto3' in sys.modules:
+			return self.__boto3_delete__(key)
+
 		remote_key = Key(self.bucket)
 		remote_key.key = key
 		remote_key.delete()
 
 	def list(self, prefix=None):
 		""" See :py:meth:`RigorS3Client.list` """
+		if 'boto3' in sys.modules:
+			return self.__boto3_list__(prefix)
+
 		if prefix is None:
 			prefix = ""
 		return self.bucket.list(prefix=prefix)
+
+	def __boto3_get__(self, key, local_file=None):
+		""" See :py:meth:`RigorS3Client.get` 
+			This implementation is speicfic to Boto3
+		"""
+		s3_obj = BytesIO()
+		try:
+			if local_file is None:
+				self.bucket.download_fileobj(key, s3_obj)
+				s3_obj.seek(0)
+			else:
+				self.bucket.download_file(key, local_file)
+				return True
+		except botocore.exceptions.ClientError as e:
+			return None
+
+		return s3_obj
+
+	def __boto3_put__(self, key, data):
+		""" See :py:meth:`RigorS3Client.put` 
+			This implementation is specific to Boto3
+		"""
+		if hasattr(data, 'read'):
+			self.bucket.upload_fileobj(data, key)
+		else:
+			self.bucket.upload_file(data, key)
+
+	def __boto3_delete__(self, key):
+		""" See :py:meth:`RigorS3Client.delete` 
+			This implementation is specific to Boto3
+		"""
+		delObject = {'Objects': [{'Key': key}]}
+		self.bucket.delete_objects(Delete=delObject)
+
+	def __boto3_list__(self, prefix=None):
+		""" See :py:meth:`RigorS3Client.list` 
+			This implementation is specific to Boto3
+		"""
+		if prefix is None:
+			return self.bucket.objects.all()
+		else:
+			return self.bucket.objects.filter(Prefix=prefix)
+
+
+
+
 
 DefaultS3Client = BotoS3Client
